pyproject.toml
[project]
name = "salesforce-docs-rag"
version = "0.1.0"
description = "Production-grade RAG system over public Salesforce documentation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12",
  "fastapi>=0.111",
  "httpx>=0.27",
  "lxml>=5.2",
  "openai>=1.40",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "scrapy>=2.11",
  "tenacity>=8.2",
  "uvicorn[standard]>=0.30",
  "weaviate-client>=4.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
]
airflow = [
  "apache-airflow>=2.9,<3.0",
]

[project.scripts]
sfdc-rag-crawl = "salesforce_docs_rag.ingestion.run_crawl:main"
sfdc-rag-index = "salesforce_docs_rag.ingestion.index_documents:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/salesforce_docs_rag"]
artifacts = ["scrapy_project"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
